+++ consys-auto-consistency-level-inferring/src/main/java/de/tuda/stg/consys/invariants/subset/visitors/FormulaGenerator.java
@@ -1,20 +1,27 @@
-package subset.visitors;
+package de.tuda.stg.consys.invariants.subset.visitors;
 
 import com.microsoft.z3.*;
+import de.tuda.stg.consys.invariants.exceptions.UnsupportedJMLExpression;
+import de.tuda.stg.consys.invariants.exceptions.WrongJMLArgumentsExpression;
+import de.tuda.stg.consys.invariants.subset.Z3Checker;
+import de.tuda.stg.consys.invariants.subset.z3_model.InternalArray;
+import de.tuda.stg.consys.invariants.subset.z3_model.InternalScope;
+import de.tuda.stg.consys.invariants.subset.z3_model.InternalVar;
 import org.eclipse.jdt.internal.compiler.ast.*;
 import org.jmlspecs.jml4.ast.*;
-import subset.Z3Checker;
-import subset.z3_model.InternalArray;
-import subset.z3_model.InternalScope;
-import subset.z3_model.InternalVar;
 
-import java.util.*;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
  * This visitor class is used to translate JML expressions, local variable declarations and
  * assignable clauses to Z3 expressions
  */
+@SuppressWarnings("rawtypes")
 public class FormulaGenerator {
   Context context = Z3Checker.context;
   /**
@@ -30,27 +37,20 @@ public class FormulaGenerator {
    * @return the corresponding Z3 expression
    */
   public Expr visitExpression(Expression expression, InternalScope scope) {
-    // "\result" is the result reference
-    if (expression instanceof JmlResultReference) return scope.getCurrentReturnVariable();
 
-    // literals can be translated directly
-    if (expression instanceof IntLiteral) return context.mkInt(((IntLiteral) expression).value);
-    if (expression instanceof DoubleLiteral) {
-      double value = ((DoubleLiteral) expression).constant.doubleValue();
-      return context.mkFPToReal(context.mkFP(value, context.mkFPSortDouble()));
+    // literal expression: 10, -5.6, true, ...
+    if (expression instanceof Literal) {
+      return visitLiteralExpression((Literal) expression, scope);
     }
-    if (expression instanceof TrueLiteral) return context.mkTrue();
-    if (expression instanceof FalseLiteral) return context.mkFalse();
 
-    // expressions like addition, modulo, and, or, ...
-    if (expression instanceof BinaryExpression) {
-      return visitBinaryExpression((BinaryExpression) expression, scope);
-    }
 
     // one reference of a variable: "a"
     if (expression instanceof JmlSingleNameReference) {
       return visitJmlSingleReference((JmlSingleNameReference) expression, scope);
     }
+
+    // \result is the result reference
+    if (expression instanceof JmlResultReference) return scope.getCurrentReturnVariable();
 
     // "\old(...)"
     if (expression instanceof JmlOldExpression) {
@@ -76,7 +76,56 @@ public class FormulaGenerator {
     if (expression instanceof JmlMessageSend) {
       return visitJmlMessageSend((JmlMessageSend) expression, scope);
     }
-    return null;
+
+    // expressions with operators: a + b, a ? b : c, !a ...
+    if (expression instanceof OperatorExpression) {
+      return visitOperatorExpression((OperatorExpression) expression, scope);
+    }
+
+    throw new UnsupportedJMLExpression(expression);
+  }
+
+  public Expr visitLiteralExpression(org.eclipse.jdt.internal.compiler.ast.Literal literalExpression, InternalScope scope) {
+    // literals can be translated directly
+    if (literalExpression instanceof IntLiteral)
+      return context.mkInt(((IntLiteral) literalExpression).value);
+
+    if (literalExpression instanceof DoubleLiteral) {
+      double value = ((DoubleLiteral) literalExpression).constant.doubleValue();
+      return context.mkFPToReal(context.mkFP(value, context.mkFPSortDouble()));
+    }
+
+    if (literalExpression instanceof TrueLiteral)
+      return context.mkTrue();
+
+    if (literalExpression instanceof FalseLiteral)
+      return context.mkFalse();
+
+    throw new UnsupportedJMLExpression(literalExpression);
+  }
+
+  public Expr visitOperatorExpression(OperatorExpression operatorExpression, InternalScope scope) {
+    // !a ...
+    if (operatorExpression instanceof UnaryExpression) {
+      return visitUnaryExpression((UnaryExpression) operatorExpression, scope);
+    }
+
+    // expressions like addition, modulo, and, or, ...
+    if (operatorExpression instanceof BinaryExpression) {
+      return visitBinaryExpression((BinaryExpression) operatorExpression, scope);
+    }
+
+    if (operatorExpression instanceof ConditionalExpression) {
+      return visitConditionalExpression((ConditionalExpression) operatorExpression, scope);
+    }
+
+    throw new UnsupportedJMLExpression(operatorExpression);
+  }
+
+  public Expr visitUnaryExpression(UnaryExpression unaryExpression, InternalScope scope) {
+    Expr expr = visitExpression(unaryExpression.expression, scope);
+
+    throw new UnsupportedJMLExpression(unaryExpression);
   }
 
   /**
@@ -92,17 +141,24 @@ public class FormulaGenerator {
 
     // get operator value and construct corresponding Z3 expression
     String s = binaryExpression.operatorToString();
-    if ((s.equals("&&") || s.equals("&"))
-        && left instanceof BoolExpr
-        && right instanceof BoolExpr) {
-      return context.mkAnd((BoolExpr) left, (BoolExpr) right);
+
+
+
+    if (s.equals("&&") || s.equals("&")) {
+      if (left instanceof BoolExpr && right instanceof BoolExpr) {
+        return context.mkAnd((BoolExpr) left, (BoolExpr) right);
+      }
+      throw new WrongJMLArgumentsExpression(binaryExpression);
     }
 
-    if ((s.equals("||") || s.equals("|"))
-        && left instanceof BoolExpr
-        && right instanceof BoolExpr) {
-      return context.mkOr((BoolExpr) left, (BoolExpr) right);
+    if (s.equals("||") || s.equals("|")) {
+      if (left instanceof BoolExpr && right instanceof BoolExpr) {
+        return context.mkOr((BoolExpr) left, (BoolExpr) right);
+      }
+      throw new WrongJMLArgumentsExpression(binaryExpression);
     }
+
+    //TODO: Add WrongJMLArgumentsExpression to all cases
     if (s.equals("<") && left instanceof ArithExpr && right instanceof ArithExpr) {
       return context.mkLt((ArithExpr) left, (ArithExpr) right);
     }
@@ -164,7 +220,8 @@ public class FormulaGenerator {
 
     if (left != null && right != null && (s.equals("<=!=>") || s.equals("!=")))
       return context.mkNot(context.mkEq(left, right));
-    return null;
+
+    throw new UnsupportedJMLExpression(binaryExpression);
   }
 
   /**
@@ -183,7 +240,8 @@ public class FormulaGenerator {
       return scope.getClassVariable(variableName).getNewValue();
     else if (scope.getLocalVariables().containsKey(variableName))
       return scope.getLocalVariable(variableName);
-    return null;
+
+    throw new WrongJMLArgumentsExpression(jmlSingleNameReference);
   }
 
   /**
@@ -218,6 +276,21 @@ public class FormulaGenerator {
 
     // substitute all occurrences of poststate variables with the prestate ones
     return subExpr.substitute(newVars, oldVars);
+  }
+
+  public Expr visitConditionalExpression(ConditionalExpression conditionalExpression, InternalScope scope) {
+    Expr cond = visitExpression(conditionalExpression.condition, scope);
+    Expr thenBranch = visitExpression(conditionalExpression.valueIfTrue, scope);
+    Expr elseBranch = visitExpression(conditionalExpression.valueIfFalse, scope);
+
+    if (cond instanceof BoolExpr) {
+      BoolExpr condBool = (BoolExpr) cond;
+
+      return context.mkITE(condBool, thenBranch, elseBranch);
+    }
+
+    throw new WrongJMLArgumentsExpression(conditionalExpression);
+
   }
 
   /**
@@ -255,7 +328,8 @@ public class FormulaGenerator {
         return context.mkString(className);
       }
     }
-    return null;
+
+    throw new WrongJMLArgumentsExpression(jmlQualifiedNameReference);
   }
 
   /**
@@ -272,7 +346,7 @@ public class FormulaGenerator {
       return context.mkSelect((ArrayExpr) array, index);
     }
 
-    return null;
+    throw new WrongJMLArgumentsExpression(jmlArrayReference);
   }
 
   /**
@@ -284,6 +358,7 @@ public class FormulaGenerator {
    */
   public Expr visitJmlQuantifiedExpression(
       JmlQuantifiedExpression jmlQuantifiedExpression, InternalScope scope) {
+
     // boundVariables declaration: introduce new local scope
     Map<String, Expr> newLocalVars = new HashMap<>(scope.getLocalVariables());
     Expr[] boundConstants = new Expr[jmlQuantifiedExpression.boundVariables.length];
@@ -308,52 +383,66 @@ public class FormulaGenerator {
     Expr rangeExpr = visitExpression(jmlQuantifiedExpression.range, newLocalScope);
     Expr bodyExpr = visitExpression(jmlQuantifiedExpression.body, newLocalScope);
 
+    //quantifier operator as string
+    String quantifier = jmlQuantifiedExpression.quantifier.lexeme;
+
     // this applies to \forall and \exists expressions
-    if (rangeExpr instanceof BoolExpr && bodyExpr instanceof BoolExpr) {
-      // range ==> body
-      BoolExpr finalBodyExpr = context.mkImplies((BoolExpr) rangeExpr, (BoolExpr) bodyExpr);
+    if (quantifier.equals(JmlQuantifier.EXISTS) || quantifier.equals(JmlQuantifier.FORALL)) {
 
-      // quantifier: forall or exists
-      Expr quantifiedExpr =
-          context.mkQuantifier(
-              jmlQuantifiedExpression.quantifier.lexeme.equalsIgnoreCase("\\forall"),
-              boundConstants,
-              finalBodyExpr,
-              1,
-              null,
-              null,
-              context.mkSymbol("Q_" + quantifierCounter),
-              context.mkSymbol("Sk_" + quantifierCounter));
+      if (rangeExpr instanceof  BoolExpr && bodyExpr instanceof BoolExpr) {
+        // range ==> body
+        BoolExpr finalBodyExpr = context.mkImplies((BoolExpr) rangeExpr, (BoolExpr) bodyExpr);
 
-      // increment for future quantifiers
-      quantifierCounter++;
+        boolean isForall = quantifier.equals(JmlQuantifier.FORALL);
 
-      return quantifiedExpr;
+        // quantifier: forall or exists
+        Expr quantifiedExpr =
+                context.mkQuantifier(
+                        isForall, //decide whether to create forall or exists quantifier
+                        boundConstants,
+                        finalBodyExpr,
+                        1,
+                        null,
+                        null,
+                        context.mkSymbol("Q_" + quantifierCounter),
+                        context.mkSymbol("Sk_" + quantifierCounter));
+
+        // increment for future quantifiers
+        quantifierCounter++;
+
+        return quantifiedExpr;
+      }
+
+      throw new WrongJMLArgumentsExpression(jmlQuantifiedExpression);
     }
 
     // this applies to \sum expressions
-    if (rangeExpr instanceof BoolExpr
-        && bodyExpr instanceof ArithExpr
-        && jmlQuantifiedExpression.quantifier.lexeme.equalsIgnoreCase("\\sum")) {
-      // \sum int b; b>=0 && b<10; \old(incs[b])
-      // only support whole array expressions as of now
-      // get array and internalArray instance and call combine with ADD
-      if (bodyExpr.isSelect()) {
-        ArrayExpr selectedArray = (ArrayExpr) bodyExpr.getArgs()[0];
-        // name is something like old_array, new_array or other_array
-        String name =
-            selectedArray.toString().substring(selectedArray.toString().lastIndexOf("_") + 1);
-        if (scope.getClassVariable(name) instanceof InternalArray) {
-          InternalArray arr = (InternalArray) scope.getClassVariable(name);
-          return arr.combineValues(
-              InternalArray.Combiner.ADDITION,
-              selectedArray,
-              ((ArraySort) arr.getSort()).getRange().getSortKind());
+    if (quantifier.equals(JmlQuantifier.SUM)) {
+      if (rangeExpr instanceof BoolExpr && bodyExpr instanceof ArithExpr) {
+
+
+        // \sum int b; b>=0 && b<10; \old(incs[b])
+        // only support whole array expressions as of now
+        // get array and internalArray instance and call combine with ADD
+        if (bodyExpr.isSelect()) {
+          ArrayExpr selectedArray = (ArrayExpr) bodyExpr.getArgs()[0];
+          // name is something like old_array, new_array or other_array
+          String name =
+                  selectedArray.toString().substring(selectedArray.toString().lastIndexOf("_") + 1);
+          if (scope.getClassVariable(name) instanceof InternalArray) {
+            InternalArray arr = (InternalArray) scope.getClassVariable(name);
+            return arr.combineValues(
+                    InternalArray.Combiner.ADDITION,
+                    selectedArray,
+                    ((ArraySort) arr.getSort()).getRange().getSortKind());
+          }
         }
       }
+
+      throw new WrongJMLArgumentsExpression(jmlQuantifiedExpression);
     }
 
-    return null;
+    throw new UnsupportedJMLExpression(jmlQuantifiedExpression);
   }
 
   /**
@@ -387,7 +476,7 @@ public class FormulaGenerator {
       }
     }
 
-    return null;
+    throw new WrongJMLArgumentsExpression(jmlMessageSend);
   }
 
   /*
@@ -410,7 +499,7 @@ public class FormulaGenerator {
       return context.mkConst(String.valueOf(localDeclaration.name) + quantifierCounter, type);
     }
 
-    return null;
+    throw new IllegalArgumentException(localDeclaration.toString());
   }
 
   /**
@@ -523,6 +612,6 @@ public class FormulaGenerator {
       return resultingPostCondition;
     }
 
-    return null;
+    throw new IllegalArgumentException(jmlAssignableClause.toString());
   }
 }