+++ consys-auto-consistency-level-inferring/src/main/java/de/tuda/stg/consys/invariants/subset/z3_model/InternalMethod.java
@@ -1,10 +1,10 @@
-package subset.z3_model;
+package de.tuda.stg.consys.invariants.subset.z3_model;
 
 import com.microsoft.z3.BoolExpr;
 import com.microsoft.z3.Context;
 import com.microsoft.z3.Expr;
 import com.microsoft.z3.Sort;
-import subset.Z3Checker;
+import de.tuda.stg.consys.invariants.subset.Z3Checker;
 
 import java.util.HashMap;
 import java.util.Map;