+++ consys-auto-consistency-level-inferring/src/test/java/TestOutput.java
@@ -1,11 +1,11 @@
-import org.checkerframework.checker.signature.qual.DotSeparatedIdentifiersOrPrimitiveType;
+import de.tuda.stg.consys.invariants.Main;
 import org.eclipse.jdt.core.compiler.CompilationProgress;
 import org.eclipse.jdt.internal.compiler.ast.TypeDeclaration;
 import org.jmlspecs.jml4.ast.JmlTypeDeclaration;
 import org.junit.jupiter.api.Test;
-import subset.Z3Checker;
-import subset.visitors.ModelGenerator;
-import subset.z3_model.InternalClass;
+import de.tuda.stg.consys.invariants.subset.Z3Checker;
+import de.tuda.stg.consys.invariants.subset.visitors.ModelGenerator;
+import de.tuda.stg.consys.invariants.subset.z3_model.InternalClass;
 
 import java.io.PrintWriter;
 import java.util.List;