+++ consys-auto-consistency-level-inferring/src/test/java/TestZ3Analysis.java
@@ -3,10 +3,10 @@ import org.eclipse.jdt.internal.compiler.ast.TypeDeclaration;
 import org.jmlspecs.jml4.ast.JmlTypeDeclaration;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import subset.Z3Checker;
-import subset.visitors.ModelGenerator;
-import subset.z3_model.InternalClass;
-import subset.z3_model.InternalMethod;
+import de.tuda.stg.consys.invariants.subset.Z3Checker;
+import de.tuda.stg.consys.invariants.subset.visitors.ModelGenerator;
+import de.tuda.stg.consys.invariants.subset.z3_model.InternalClass;
+import de.tuda.stg.consys.invariants.subset.z3_model.InternalMethod;
 
 import java.io.PrintWriter;
 import java.util.List;